### 2016-05-04 #11 @whedon
Submission: carl: a likelihood-free inference toolbox
**Submitting author:** @glouppe (<a href="http://orcid.org/0000-0002-2082-3106">Gilles Louppe</a>)
**Repository:** <a href="https://github.com/diana-hep/carl" target="_blank">https://github.com/diana-hep/carl</a>
**Version:** v0.2
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.47798" target="_blank" >10.5281/zenodo.47798</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.47798">software archive</a> resolve?

- ~~[ ] **Performance:** Have the performance claims of the software been confirmed?~~

- [x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? diana-hep/carl#53
- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support diana-hep/carl#53

- [x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? diana-hep/carl#53
