### 2016-05-04 #11 @cranmer
Thanks @whedon would be nice if arxiv showed up in bib. should we edit bib or change style for latex?
