### 2016-05-04 #11 @arfon
@cranmer This is the command that we're running to compile from Markdown to PDF: https://github.com/openjournals/whedon/blob/master/lib/whedon/processor.rb#L64 and this is the latex template: https://github.com/openjournals/whedon/blob/master/resources/latex.template

Any suggestions/modifications that would make arXiv show up very welcome 😄
