### 2016-05-04 #11 @betatim
As a reviewer should I do things beyond what is listed in the initial comment on this PR or is that it? If it is only those questions I can review this and use this opportunity to declare that I am a good friend of @glouppe but don't think this creates a conflict of interest.
