### 2016-05-04 #11 @arfon
@betatim - basically we just need you to work through the checklist  under `Reviewer questions`. If anything isn't clear please shout.

Also, there are some additional notes here (http://joss.theoj.org/about#reviewer_guidelines) but the review checklist should represent these guidelines.
