### 2016-05-05 #11 @betatim
UI question: how do I tick off things from the list? I think technically it is impossible to organise for me to be able to edit @whedon's comment. Had a scan around openjournals/whedon to try and find the raw MD but couldn't find it sad 🐼.
