### 2016-05-05 #11 @betatim
- [ ] **Repository:** Is the source code for this software available at the
- [ ] **Version:** Does the release version given match the GitHub release ()?
- [ ] **Archive:** Does the software archive resolve?

- [ ] **Performance:** Have the performance claims of the software been confirmed?

- [ ] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
