### 2016-05-05 #11 @betatim
A few "trivial" comments are included inline. I haven't yet tried to answer the "functionality" questions.

re: statement of need in the PDF. It states "support inference in the likelihood-free setup, including density ratio estimation algorithms, parameterized supervised learning and calibration procedures". This could be improved by stating a few concrete problems, that your average researcher can recognise as problems they have, to which it is the solution. Right now I think you have to already know that `carl` can solve your problem to realise that it solves your problem.

The first paragraph of the PDF is a bit longer than what is in the `README` so I would copy/sync them once a few specific problems are added to it.

Otherwise everything looked good to me.
