### 2016-05-05 2a92a97
DOC: address @betatim comments
