### 2016-05-05 #11 @glouppe
Thanks Tim! Changes proposed are https://github.com/diana-hep/carl/pull/53

Note: this back and forth between two github repos is not the most convenient :s
