### 2016-05-05 #11 @arfon
> UI question: how do I tick off things from the list? I think technically it is impossible to organise for me to be able to edit @whedon's comment. Had a scan around openjournals/whedon to try and find the raw MD but couldn't find it sad 🐼.

@betatim - this is great point (that I didn't consider until now) - you need to be able to edit the issue too :-) I've added you to the `joss-reviewers` team which means you should be able to update the original issue.

Also, thanks for the rapid review :zap:
