### 2016-05-05 #11 @betatim
I've moved everything to the first comment. Most of the missing ☑️  are being dealt with in diana-hep/carl#53. Then there is the missing DOI in the generated PDF, do you take care of that @arfon?
