### 2016-05-05 #11 @arfon
> Then there is the missing DOI in the generated PDF, do you take care of that @arfon?

Yep, tracking that in https://github.com/openjournals/joss/issues/56
