### 2016-05-05 8322d60
DOC: added CONTRIBUTING.md
