+++ CONTRIBUTING.md
@@ -1,0 +1,15 @@
+# Contributing
+
+Carl is an open-source project and contributions of all kinds
+are welcome.
+
+Code contributions should follow these guidelines:
+
+* all changes by pull request (PR);
+* a PR solves one problem (do not mix problems together in one PR) with the
+  minimal set of changes;
+* describe why you are proposing the changes you are proposing;
+* new code needs to come with a test;
+* no merging if travis is red.
+
+These are not hard rules to be enforced, but guidelines.
