### 2016-05-06 #11 @cranmer
I don't see what the default bib style is in the wehedon resources:

https://github.com/openjournals/whedon/blob/71a1a35a21cce942378229892976ddc4c26e57e8/resources/latex.template#L98

but it looks like `\bibliographystyle{utphys}` might give a nice format for the arxiv submission

http://tex.stackexchange.com/questions/243636/mendeley-and-arxiv-citation-style
