### 2016-05-08 e8456f4
Towards dev setup instructions
