### 2016-05-08 589ba33
How to run the tests
