### 2016-05-08 #11 @betatim
Installing works, automated tests pass, and ran some of the examples.

There are no performance claims (as in runs faster than X) so I don't think that reviewer question applies.

Only outstanding issue is rendering of references in the pdf.
