### 2016-05-08 #11 @arfon
@betatim - awesome thanks.

> Only outstanding issue is rendering of references in the pdf.

@cranmer @betatim how do you feel about this rendering? [paper.pdf](https://github.com/openjournals/joss-reviews/files/253947/paper.pdf)

I've not managed to get the `utphys` biblio-style working but have managed to format the reference more generically to get the arXiv URL included:

```
@article{Cranmer:2015-llr,
      author         = "Cranmer, Kyle and Pavez, Juan and Louppe, Gilles",
      title          = "{Approximating Likelihood Ratios with Calibrated
                        Discriminative  Classifiers}",
      eprinttype = {arxiv},
      version     = {2},
      eprint      = {http://arxiv.org/abs/1506.02169v2},
      url         = {http://arxiv.org/abs/1506.02169v2},
      date        = {2016-03-18},
}

```
