### 2016-05-08 #11 @betatim
LGTM. I use the arXiv ID/DOI as a simple way to find the article in google, if you directly link me there even better (appeals to the efficient German in me)
