### 2016-05-09 f5f65f5
DOC: improve README
