### 2016-05-10 #11 @arfon
Ok then I think we're done here. Congrats @glouppe and @cranmer on our first accepted paper!

I'll finish up the editorial pieces this evening.

🚀
