### 2016-05-11 #11 @arfon
> Actually, where do we click to toggle the badge from under review to published?

@betatim - I have to do this (for now).
