### 2016-05-12 #11 @danielskatz
It seems like the reference is still broken - there's no DOI or URL or way to know where the reference was published or is available
