### 2016-05-12 #11 @arfon
> It seems like the reference is still broken - there's no DOI or URL or way to know where the reference was published or is available

Which reference sorry?
