### 2016-05-12 #11 @betatim
Compare https://github.com/openjournals/joss-papers/blob/master/joss.00011/10.21105.joss.00011.pdf (from the journal page) to the last PDF you posted in this issue. The only reference in the PDF has a link to the arxiv in the version you posted here but not in the journal version
