### 2016-05-13 #11 @arfon
OK, this is fixed in the online PDF now. The PR to fix the bibtex is here: https://github.com/diana-hep/carl/pull/56
